refactor:move can detail to directory
