+++ frontend/src/pages/cans/detail/canDetailSlice.js
@@ -1,6 +1,6 @@
 import { createSlice } from "@reduxjs/toolkit";
-import constants from "../constants";
-import { getCurrentFiscalYear } from "../helpers/utils";
+import constants from "../../../constants";
+import { getCurrentFiscalYear } from "../../../helpers/utils";
 
 const initialState = {
     can: {},
