+++ frontend/src/pages/cans/detail/getCan.js
@@ -1,4 +1,4 @@
-import ApplicationContext from "../applicationContext/ApplicationContext";
+import ApplicationContext from "../../../applicationContext/ApplicationContext";
 
 export const getCan = async (id) => {
     const api_version = ApplicationContext.get().helpers().backEndConfig.apiVersion;
